+++ scout/server/blueprints/institutes/controllers.py
@@ -99,7 +99,6 @@ def update_institute_settings(store, institute_obj, form):
     gene_panels = {}
     group_abbreviations = []
     cohorts = []
-    loqusdb_id = []
 
     for email in form.getlist("sanger_emails"):
         sanger_recipients.append(email.strip())
@@ -124,9 +123,6 @@ def update_institute_settings(store, institute_obj, form):
     for cohort in form.getlist("cohorts"):
         cohorts.append(cohort.strip())
 
-    if form.get("loqusdb_id"):
-        loqusdb_id.append(form.get("loqusdb_id"))
-
     updated_institute = store.update_institute(
         internal_id=institute_obj["_id"],
         sanger_recipients=sanger_recipients,
@@ -140,6 +136,8 @@ def update_institute_settings(store, institute_obj, form):
         sharing_institutes=sharing_institutes,
         cohorts=cohorts,
         loqusdb_id=form.get("loqusdb_id"),
+        alamut_institute=form.get("alamut_institute"),
+        alamut_key=form.get("alamut_key"),
     )
     return updated_institute